test: isolate real I/O and native compatibility probes
