+++ vitest.compat.config.ts
@@ -1,0 +1,11 @@
+import { defineConfig } from 'vitest/config';
+import config from './vitest.config.ts';
+
+export default defineConfig({
+  ...config,
+  test: {
+    ...config.test,
+    include: ['packages/*/src/**/*.compat.test.ts'],
+    exclude: [],
+  },
+});
